Edits

+++ Model/Mail.java
@@ -36,7 +36,7 @@ public class Mail implements Comparable<Mail> {
     /**
      * Just use the getDistance() method of the Location class and return it.
      */
-    public int getShortestPath() {
+    public double getShortestPath() {
         return origin.getShortestPath(destination);
     }
